Updated to new session model

+++ libs/SecurityDecorators.py
@@ -31,21 +31,21 @@ def authenticated(method):
 
     @functools.wraps(method)
     def wrapper(self, *args, **kwargs):
-        if self.session is not None:
-            if self.session.ip_address == self.request.remote_ip:
+        if self.session is not None and 'remote_ip' in self.session:
+            if self.session['remote_ip'] == self.request.remote_ip:
                 if not self.get_current_user().locked: 
                     return method(self, *args, **kwargs)
                 else:
                     self.session.delete()
                     self.clear_all_cookies()
-                    self.redirect('/403?locked=true')
+                    self.redirect('/403')
             else:
                 logging.warn("Session hijack attempt from %s?" % (
                     self.request.remote_ip,
                 ))
-                self.redirect(self.application.settings['login_url'])
+                self.redirect('/403')
         else:
-            self.redirect(self.application.settings['login_url'])
+            self.redirect('/login')
     return wrapper
 
 
@@ -90,7 +90,6 @@ def async(method):
     def __async__(*args, **kwargs):
         worker = Thread(target=method, args=args, kwargs=kwargs)
         worker.start()
-    #return __async__
 
 
 def debug(method):
